refactor(Participant): Player와 Dealer를 Participant로 추상화한다
